Safari-609.1.11.1/LayoutTests/inspector/css/stylesheet-events-imports-expected.txt

Test for CSS.styleSheetAdded and CSS.styleSheetRemoved events with @import rules.


== Running test suite: CSS.StyleSheetEvents.Imports
-- Running test case: CheckStyleSheets
PASS: Should be 3 stylesheets.
inspector/css/stylesheet-events-imports.html
inspector/css/resources/import-level-1.css
inspector/css/resources/import-level-2.css

-- Running test case: DisablePropagatesThroughImports
StyleSheetRemoved: inspector/css/resources/import-level-1.css
StyleSheetRemoved: inspector/css/resources/import-level-2.css
StyleSheetRemoved: inspector/css/stylesheet-events-imports.html

-- Running test case: EnablePropagatesThroughImports
StyleSheetAdded: inspector/css/resources/import-level-1.css
StyleSheetAdded: inspector/css/resources/import-level-2.css
StyleSheetAdded: inspector/css/stylesheet-events-imports.html

-- Running test case: DeleteImportRemovesStyleSheet
StyleSheetRemoved: inspector/css/resources/import-level-1.css
StyleSheetRemoved: inspector/css/resources/import-level-2.css
